# docs/postman_OpenApiV3_response.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Car-Rest-Service API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/cars/{objectId}": {
      "get": {
        "tags": [
          "cars"
        ],
        "summary": "Get car by objectId",
        "description": "Retrieve a specific car by its unique objectId",
        "operationId": "getCarByObjectId",
        "parameters": [
          {
            "name": "objectId",
            "in": "path",
            "description": "Unique identifier of the car",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "ZRgPP9dBMm"
          }
        ],
        "responses": {
          "200": {
            "description": "Car found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid objectId format"
          },
          "404": {
            "description": "Car not found with provided objectId"
          }
        }
      },
      "put": {
        "tags": [
          "cars"
        ],
        "summary": "Update car by objectId",
        "description": "Update an existing car record. Requires authentification.",
        "operationId": "updateCar",
        "parameters": [
          {
            "name": "objectId",
            "in": "path",
            "description": "Unique identifier of the car",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "ZRgPP9dBMm"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Car updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or objectId format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Car not found with provided objectId"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "cars"
        ],
        "summary": "Delete car by objectId",
        "description": "Delete a specific car record. Requires authentification.",
        "operationId": "deleteCarByObjectId",
        "parameters": [
          {
            "name": "objectId",
            "in": "path",
            "description": "Unique identifier of the car",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "ZRgPP9dBMm"
          }
        ],
        "responses": {
          "204": {
            "description": "Car deleted successfully"
          },
          "400": {
            "description": "Invalid objectId format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Car not found with provided objectId"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/cars": {
      "get": {
        "tags": [
          "cars"
        ],
        "summary": "Get all cars with filtering and pagination",
        "description": "Retrieve a paginated list of cars with optional filtering by make, model, year, etc.",
        "operationId": "getAllCars",
        "parameters": [
          {
            "name": "carFilterDto",
            "in": "query",
            "description": "Filter criteria for cars",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CarFilterDto"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "Pagination information",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cars retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter parameters"
          }
        }
      },
      "post": {
        "tags": [
          "cars"
        ],
        "summary": "Create a new car",
        "description": "Create a new car record. Requires authentification.",
        "operationId": "createCar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Car created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "409": {
            "description": "Car already exists"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "MakeDto": {
        "type": "object",
        "properties": {
          "makeId": {
            "type": "integer",
            "format": "int32"
          },
          "make": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "make"
        ]
      },
      "ModelDto": {
        "type": "object",
        "properties": {
          "modelId": {
            "type": "integer",
            "format": "int32"
          },
          "make": {
            "$ref": "#/components/schemas/MakeDto"
          },
          "model": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "make",
          "model"
        ]
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "category"
        ]
      },
      "CarDto": {
        "type": "object",
        "properties": {
          "objectId": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "model": {
            "$ref": "#/components/schemas/ModelDto"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "model",
          "objectId",
          "year"
        ]
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {}
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "paged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
      "CarFilterDto": {
        "type": "object",
        "properties": {
          "manufacturer": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "minYear": {
            "type": "integer",
            "format": "int32"
          },
          "maxYear": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}